refactor: 매핑 함수 위치 변경
