--- app/src/main/java/woowacourse/shopping/mapper/ProductMapper.kt
@@ -1,14 +1,0 @@
-package woowacourse.shopping.mapper
-
-import woowacourse.shopping.domain.model.Product
-import woowacourse.shopping.presentation.product.catalog.ProductUiModel
-
-fun Product.toUiModel() =
-    ProductUiModel(
-        id = this.id,
-        imageUrl = this.imageUrl,
-        name = this.name,
-        price = this.price,
-        category = this.category,
-        quantity = 0,
-    )
