--- app/src/main/java/woowacourse/shopping/mapper/ViewedItemMapper.kt
@@ -1,13 +1,0 @@
-package woowacourse.shopping.mapper
-
-import woowacourse.shopping.data.model.ViewedItem
-import woowacourse.shopping.presentation.product.catalog.ProductUiModel
-
-fun ViewedItem.toUiModel() =
-    ProductUiModel(
-        id = this.id,
-        imageUrl = this.imageUrl,
-        name = this.name,
-        price = this.price,
-        category = this.category,
-    )
