+++ app/src/main/java/woowacourse/shopping/presentation/product/catalog/ProductUiModel.kt
@@ -2,6 +2,7 @@ package woowacourse.shopping.presentation.product.catalog
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import woowacourse.shopping.domain.model.Product
 
 @Parcelize
 data class ProductUiModel(
@@ -14,3 +15,22 @@ data class ProductUiModel(
     val isExpanded: Boolean = false,
     val isChecked: Boolean = false,
 ) : Parcelable
+
+fun Product.toUiModel() =
+    ProductUiModel(
+        id = this.id,
+        imageUrl = this.imageUrl,
+        name = this.name,
+        price = this.price,
+        category = this.category,
+        quantity = 0,
+    )
+
+fun ProductUiModel.toDomain() =
+    Product(
+        id = this.id,
+        name = this.name,
+        price = this.price,
+        imageUrl = this.imageUrl,
+        category = this.category,
+    )
